"Softening" expressions: silently return undef on impending errors
Things like subselecting an undefined value used to throw an exception.
In the context of template expressions, however, it should be perfectly
legal to not define (or overdefine) data.

One syntax change emerged from this:
'hashref.func()' now looks up 'func' in the hashref, and if found,
executes that function with *only* the given arguments. To determine
the key via a function, use 'hashref.(func())'.